+++ src/builtins/built_in_export.c
@@ -6,7 +6,7 @@
 /*   By: danslav1e <danslav1e@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/15 01:40:40 by danslav1e         #+#    #+#             */
-/*   Updated: 2025/11/15 02:18:32 by danslav1e        ###   ########.fr       */
+/*   Updated: 2025/11/20 01:36:20 by danslav1e        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -81,7 +81,7 @@ int	set_env_var(t_shell_state *state, char *var)
  */
 int	handle_export_args(t_node *node, t_shell_state *state)
 {
-	int		i;
+	size_t		i;
 	int		exit_code;
 	char	*arg;
