+++ src/builtins/built_in_unset.c
@@ -6,7 +6,7 @@
 /*   By: danslav1e <danslav1e@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/10 16:41:28 by danslav1e         #+#    #+#             */
-/*   Updated: 2025/11/15 02:38:51 by danslav1e        ###   ########.fr       */
+/*   Updated: 2025/11/20 01:35:56 by danslav1e        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ void	remove_var(int index, t_shell_state *state)
 int	built_in_unset(t_node *node, t_shell_state *state)
 {
 	int	index;
-	int	i;
+	size_t	i;
 	int	exit_code;
 
 	i = 1;
